+++ src/WebApp/PingMonitor.Web/Models/CheckResult.cs
@@ -7,7 +7,7 @@ public sealed class CheckResult
     public string AssignmentId { get; set; } = string.Empty;
     public DateTimeOffset CheckedAtUtc { get; set; }
     public bool Success { get; set; }
-    public int? RoundTripMs { get; set; }
+    public decimal? RoundTripMs { get; set; }
     public string? ErrorCode { get; set; }
     public string? ErrorMessage { get; set; }
     public DateTimeOffset ReceivedAtUtc { get; set; }
