+++ src/WebApp/PingMonitor.Web/Services/BufferedResults/BufferedCheckResult.cs
@@ -7,7 +7,7 @@ public sealed class BufferedCheckResult
     public string AssignmentId { get; init; } = string.Empty;
     public DateTimeOffset CheckedAtUtc { get; init; }
     public bool Success { get; init; }
-    public int? RoundTripMs { get; init; }
+    public decimal? RoundTripMs { get; init; }
     public string? ErrorCode { get; init; }
     public string? ErrorMessage { get; init; }
     public DateTimeOffset ReceivedAtUtc { get; init; }
