+++ src/WebApp/PingMonitor.Web/Services/Metrics/IAssignmentMetrics24hService.cs
@@ -35,9 +35,9 @@ public sealed class AssignmentMetrics24hSummary
     public long UnknownSeconds { get; init; }
     public long SuppressedSeconds { get; init; }
     public double? UptimePercent { get; init; }
-    public int? LastRttMs { get; init; }
-    public int? HighestRttMs { get; init; }
-    public int? LowestRttMs { get; init; }
+    public double? LastRttMs { get; init; }
+    public double? HighestRttMs { get; init; }
+    public double? LowestRttMs { get; init; }
     public double? AverageRttMs { get; init; }
     public double? JitterMs { get; init; }
     public DateTimeOffset? LastSuccessfulCheckUtc { get; init; }
